Move popover code to its own legacy file
