+++ draft-packages/popover/KaizenDraft/Popover/types.ts
@@ -1,0 +1,12 @@
+export type Variant =
+  | "default"
+  | "informative"
+  | "positive"
+  | "negative"
+  | "cautionary"
+
+export type Side = "top" | "bottom" // | "left" | "right" - not yet implemented
+
+export type Position = "start" | "center" | "end"
+
+export type Size = "small" | "large"
